requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
psycopg2-binary>=2.9.10,<3.0
gunicorn==23.0.0
WTForms==3.2.1
numpy==1.26.4
matplotlib==3.9.2
lxml==5.2.1
jsonpointer==2.1
leaflet==0.0.3
scikit-learn==1.6.1
protobuf==4.25.3
setuptools==75.1.0
wheel==0.44.0
python-dotenv==1.1.1

